Extended DataList refactor to View.annotations; moved DataList to model.py to avoid circular import
